requirements.txt
python-telegram-bot
openai==0.28
requests
python-dotenv